+++ make_dockerfiles.py
@@ -11,8 +11,8 @@ shas = {}
 repo_version = {}
 
 # dependencies
-versions['EMBREE_VERSION'] = '3.3.0'
-shas['EMBREE_SHA'] = 'c70e5cef5eeb88aa5c384121c8908287950d756c3eaac3f0bccea2d94c4fea3f'
+versions['EMBREE_VERSION'] = '3.5.2'
+shas['EMBREE_SHA'] = 'a2f2adf05db1ee3fb9f3116d041b36c8d272a3d3f5156ba0e042146d641db750'
 
 versions['OSU_MICROBENCHMARK_VERSION'] ='5.4.1'
 shas['OSU_MICROBENCHMARK_SHA'] ='e90cb683a01744377f77d420de401431242593d8376b25b120950266e140e83c'
@@ -20,21 +20,24 @@ shas['OSU_MICROBENCHMARK_SHA'] ='e90cb683a01744377f77d420de401431242593d8376b25b
 versions['MPI4PY_VERSION'] = '3.0.0'
 shas['MPI4PY_SHA'] = 'b457b02d85bdd9a4775a097fac5234a20397b43e073f14d9e29b6cd78c68efd7'
 
-versions['TBB_VERSION'] = '2019_U2'
-shas['TBB_SHA'] = '1245aa394a92099e23ce2f60cdd50c90eb3ddcd61d86cae010ef2f1de61f32d9'
+versions['TBB_VERSION'] = '2019_U3'
+shas['TBB_SHA'] = 'b2244147bc8159cdd8f06a38afeb42f3237d3fc822555499d7ccfbd4b86f8ece'
+
+versions['SCIPY_VERSION'] = '1.2.0'
+shas['SCIPY_SHA'] = '51a2424c8ed80e60bdb9a896806e7adaf24a58253b326fbad10f80a6d06f2214'
 
 # glotzer lab
-repo_version['fresnel']     = versions['FRESNEL_VERSION']     = 'v0.6.0'
-shas['FRESNEL_SHA'] = 'de1b18f87b5bcdd96844c143d6a9cf560df873a9f1f7eae8d6ff2eac5a1d2467'
+repo_version['fresnel']     = versions['FRESNEL_VERSION']     = 'v0.8.0'
+shas['FRESNEL_SHA'] = '0bdb1f2249717b56f351492a06a0eb3bfe9f8ed0b4f490fdd73ef4cde922de31'
 
-repo_version['freud']       = versions['FREUD_VERSION']       = 'v0.11.4'
-shas['FREUD_SHA'] = '9e54cb2f9ef2df7569ae04b5794d0372439b8667cd1ba32390496b5ddf3ad233'
+repo_version['freud']       = versions['FREUD_VERSION']       = 'v1.0.0'
+shas['FREUD_SHA'] = '71599d85be07e6ed9afc3c85ec9a5ab5eafa8c8185fd4b306eb4ae43d5e127f2'
 
-repo_version['gsd']         = versions['GSD_VERSION']         = 'v1.6.0'
-shas['GSD_SHA'] = '2d4ddacbea75b36a446a41c5df8ba16d793c5f1674e001795393931f96d0f4e9'
+repo_version['gsd']         = versions['GSD_VERSION']         = 'v1.6.1'
+shas['GSD_SHA'] = '45edc981a5899ca7fb81205a3c1a3a07d58ea955f877fdd63e2a3e15d5ead41e'
 
-repo_version['hoomd-blue']  = versions['HOOMD_VERSION']       = 'v2.4.2'
-shas['HOOMD_SHA'] = '2d46725844336c9b3cd39fcba26741042410e9d20384d5218c9272c92cccfb08'
+repo_version['hoomd-blue']  = versions['HOOMD_VERSION']       = 'v2.5.1'
+shas['HOOMD_SHA'] = '07fcc83f7fb48373fe485bf7b8ac71cb79a2a6c918da3498dd64a37f2dc2c964'
 
 repo_version['libgetar']    = versions['LIBGETAR_VERSION']    = 'v0.7.0'
 shas['LIBGETAR_SHA'] = '2a33809981b7a99c856ca60a1a7b9b1a0b3978fd8315ab3ab07b7b279a7c55e7'
@@ -42,14 +45,14 @@ shas['LIBGETAR_SHA'] = '2a33809981b7a99c856ca60a1a7b9b1a0b3978fd8315ab3ab07b7b27
 repo_version['pythia']      = versions['PYTHIA_VERSION']      = 'v0.2.4'
 shas['PYTHIA_SHA'] = 'cebc1033759f518aa4f9c41d4660c7748b646f6f6117be9e4dcb9e53ef2f0251'
 
-repo_version['rowan']       = versions['ROWAN_VERSION']       = 'v1.1.6'
-shas['ROWAN_SHA'] = '14627245b95b88e3d4358e6d9df0501eec1bcb892c71ba5829904d4728ecb9f8'
+repo_version['rowan']       = versions['ROWAN_VERSION']       = 'v1.2.0'
+shas['ROWAN_SHA'] = '68f27a6d7d1fd735a1cb0d465fd339ad137ea053fe7ccc56786464c33245dbeb'
 
 repo_version['plato']       = versions['PLATO_VERSION']       = 'v1.4.0'
 shas['PLATO_SHA'] = 'fd5b764da5fdca9e704b22629b12c83fbf482db95fe7050e2b7b7661c8e57cdf'
 
-repo_version['signac']      = versions['SIGNAC_VERSION']      = 'v0.9.4'
-shas['SIGNAC_SHA'] = '8a3c5b46d079decb9fa2d5d85628c2bd31057a44e945beba930d3b624dcb8437'
+repo_version['signac']      = versions['SIGNAC_VERSION']      = 'v1.0.0'
+shas['SIGNAC_SHA'] = '7eff681ec328a106ae9682697d2ed74565270a93410bab8583414fa756728ed0'
 
 repo_version['signac-flow'] = versions['SIGNAC_FLOW_VERSION'] = 'v0.6.4'
 shas['SIGNAC_FLOW_SHA'] = 'c261204eef08c5e954179840cdb68795f2a464c213b58e67d7b502caada4d34c'
@@ -68,6 +71,7 @@ if __name__ == '__main__':
     base_template = env.get_template('base.jinja')
     ib_mlx_template = env.get_template('ib-mlx.jinja')
     ib_hfi1_template = env.get_template('ib-hfi1.jinja')
+    ib_hfi1_stampede2_template = env.get_template('ib-hfi1-stampede2.jinja')
     openmpi_template = env.get_template('openmpi.jinja')
     mvapich2_template = env.get_template('mvapich2.jinja')
     titan_template = env.get_template('titan.jinja')
@@ -128,12 +132,13 @@ if __name__ == '__main__':
           **shas)
 
     # TODO: update cflags after switching to newer compiler
-    write('docker/stampede2/Dockerfile', [base_template, ib_hfi1_template, mvapich2_template, glotzerlab_software_template, finalize_template],
+    write('docker/stampede2/Dockerfile', [base_template, ib_hfi1_stampede2_template, mvapich2_template, glotzerlab_software_template, finalize_template],
           FROM='nvidia/cuda:9.2-devel-ubuntu16.04',
           system='stampede2',
           MVAPICH_VERSION='2.3',
           MVAPICH_PATCHLEVEL='',
           MVAPICH_SHA='01d5fb592454ddd9ecc17e91c8983b6aea0e7559aa38f410b111c8ef385b50dd',
+          MVAPICH_EXTRA_OPTS='--with-device=ch3:psm --with-ch3-rank-bits=32 --enable-cxx --enable-romio --enable-fast=O3 --enable-g=dbg',
           ENABLE_MPI='on',
           MAKEJOBS=10,
           CFLAGS='-march=knl -mmmx -msse -msse2 -msse3 -mssse3 -mcx16 -msahf -mmovbe -maes -mpclmul -mpopcnt -mabm -mfma -mbmi -mbmi2 -mavx -mavx2 -msse4.2 -msse4.1 -mlzcnt -mrtm -mhle -mrdrnd -mf16c -mfsgsbase -mrdseed -mprfchw -madx -mfxsr -mxsave -mxsaveopt -mavx512f -mavx512cd -mclflushopt -mxsavec -mxsaves -mavx512dq -mavx512bw -mclwb --param l1-cache-size=32 --param l1-cache-line-size=64 --param l2-cache-size=33792 -mtune=generic',