+++ CHANGELOG.rst
@@ -2,6 +2,42 @@
 Change log
 ==========
 
+2019.03
+-------
+
+*Software updates:*
+
+* fresnel v0.8.0
+* freud v1.0.0
+* GSD v1.6.1
+* HOOMD-blue v2.5.1
+* rowan v1.2.0
+* signac v1.0.0
+
+*Library updates:*
+
+* embree 3.5.2
+
+2019.02
+-------
+
+*Software updates:*
+
+* fresnel v0.7.1
+* HOOMD-blue v2.5.0
+* rowan v1.1.7
+* signac v0.9.5
+* [summit] scipy v1.2.0
+
+*Library updates:*
+
+* embree 3.4.0
+* TBB 2019_U3
+
+*Fixes:*
+
+* Stampede2 multi-node MPI now works
+
 2019.01
 -------